Make uid label configurable
The uid is always using the sub value from the *userinfo*, which in some applications is not the expected value.

To avoid such limitations, the uid label is now configurable by configuring the omniauth *uid_label* option to a different label that apears in the userinfo details.

The default value for *uid_label* is set to *sub* to ensure the backward compatibility is not broken.